register-server: Fix status update
Took status from object that already existed but had no status set
because status cannot be set on creation. This leads to `the object
has been modified` errors a few times until creation is successful,
leading to the creation of unnecessary Machine objects and secrets,
which was especially observable with the parallel attestation test.

Instead, set a None status first and then update using a fresh one.

Signed-off-by: Jakob Naucke <jnaucke@redhat.com>